Add files via upload
There are two kinds of time series process, namely, Moving Average(MA) process and Autoregression(AR) process. Here is how AR process is simulated in R and what the autocorrelation function(ACF) looks like for each. And go forward, the concept of partial ACF is introduced which refers to the effect of features(in this scenario, lags) without getting joint distribution involved, so called "partial out" to make all independent.